+++ src/magicdust/colors.js
@@ -1,30 +1,43 @@
 export default {
   primary: "#2c3d63",
-  red: "#ff3838",
+  red: "#B11E31",
   watermelon: "#f75961",
   blue: "#2dadba",
   orchid: "#6464dc",
   yellow: "#ffb400",
   orange: "#f57b39",
-  green: "#116138",
-  coffee: "#7e544d",
+  green: "#3CAD99",
+  coffee: "#4F3B38",
+
   bar: "#202632",
   black: "#282828",
   black2: "#353535",
   charcoal: "#2b2b2b",
+
   lead: "#393939",
   slate: "#484848",
   iron: "#686868",
   steel: "#797979",
   chrome: "#bcbcbc",
+
   grey: "#ccc",
   strokeThree: "#d9d9d9",
   strokeTwo: "#d3d3d3",
+
   strokeOne: "#ebebeb",
   shade1: "#f3f3f3",
   bg: "whitesmoke",
   shade2: "#f9f9f9",
   featherDark: "#fafafa",
   feather: "#fbfbfb",
+
   white: "#ffffff",
+
+  nbOrange: "#FF7A00",
+  nbPink: "#FF5151",
+  foBlue: "#52ADA2",
+  coBlue: "#1E2F49",
+  coGreen: "#00A962",
+  mustard: "#FFB400",
+  purple: "#908DB9",
 };