+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Timeline new structure
 - Added classes for margin & padding - 2, 4, 8, 16, 20 & 32
 - Fix hover & disabled states
+- Added nbOrange, nbPink, foBlue, coBlue, coGreen, mustard, purple colors
+- Modified red, green, coffee color codes